examples/rtl8720dn: update all remaining examples for refactored API
